Restructured the directory
