+++ example_outputs/Sequences.txt
@@ -1,0 +1,16 @@
+ATTACGTAAT,54
+ATTACGTAAT,56
+GCATAAAAAA,69
+ATTACGTAAT,85
+ATTACGTAAT,59
+GGCCATAAAA,6
+ATTACGTAAT,42
+TAATCC,40
+GGCCATAAAA,82
+GGCCATAAAA,99
+GGCCATAAAA,55
+GGCCATAAAA,18
+TAATCC,36
+ATTACGTAAT,68
+A,12
+A,90
