+++ example_outputs/motifs.txt
@@ -1,0 +1,8 @@
+> motif1
+GCATAAAAAA
+> motif2
+TAATCC
+> motif3
+GGCCATAAAA
+> motif4
+ATTACGTAAT
