+++ example_outputs/my_example.faa
@@ -1,0 +1,96 @@
+>1|0 This is a randomly generated sequence
+ATTACGTAATATTACGTAATGCATAAAAAAGGCCATAAAAATTACGTAATTAATCCGGCC
+ATAAAAGGCCATAAAATAATCCATTACGTAATGGCCATAAAAATTACGTAATGGCCATAA
+AAATTACGTAATAA
+>2|1 This is a randomly generated sequence
+GGCCATAAAAGCATAAAAAATAATCCATTACGTAATTAATCCGGCCATAAAAATTACGTA
+ATGGCCATAAAAATTACGTAATGGCCATAAAAGGCCATAAAAATTACGTAATATTACGTA
+ATATTACGTAATAA
+>3|1 This is a randomly generated sequence
+ATTACGTAATGGCCATAAAAATTACGTAATGCATAAAAAAATTACGTAATTAATCCATTA
+CGTAATGGCCATAAAAATTACGTAATGGCCATAAAAGGCCTTAAAAATTACGTAATTAAT
+CCGGCCATAAAAAA
+>4|1 This is a randomly generated sequence
+TAATCCGGCCATAAAAGGCCATAAAAGCATAAAAAAATTACGTAATGGCCATAAAAATTA
+CGTAATGGCCATAAAAATTACGTAATATTACGTAATATTACGTAATGGCCATAAAATAAT
+CCATTACGTAATAA
+>5|1 This is a randomly generated sequence
+GGCCATAAAATAATCCATTACGTAATATTACGTAATGGCCATAAAAGGCCATAAAAGGCC
+ATAAAAATTACGTAATATTACGTAATATTACGTAATATTACGTAATTAATCCGCATAAAA
+AAGGCCATAAAAAA
+>6|0 This is a randomly generated sequence
+ATTACGTAATTAATCCGGCCATAAAAATTACGTAATGGCCATAAAAGCATAAAAAAGGCC
+ATAAAAGGCCATAAAAATTACGTAATATTACGTAATTAATCCGGCCATAAAAATTACGTA
+ATATTACGTAATAA
+>7|0 This is a randomly generated sequence
+TAATCCGGCCATAAAAGGCCATAAAAGGCCATAAAAGGCCATAAAAATTACGTAATTAAT
+CCATTACGTAATGCATAAAAAAATTACGTAATATTACGTAATGGCCATAAAAATTACGTA
+ATATTACGTAATAA
+>8|0 This is a randomly generated sequence
+ATTACGTAATGGCCATAAAAATTACGTAATATTACGTAATGGCCATAAAAATTACGTAAT
+GCATAAAAAATAATCCGGCCATAAAAGGCCATAAAATAATCCATTACGTAATATTACGTA
+ATGGCCATAAAAAA
+>9|1 This is a randomly generated sequence
+ATTACGTAATGCATAAAAAAGGCCATAAAAGGCCATAAAATAATCCATTACGTAATGGCC
+ATAAAAATTACGTAATGGCCATAAAAATTACGTAATTAATCCATTACGTAATGGCCATAA
+AAATTACGTAATAA
+>10|1 This is a randomly generated sequence
+ATTACGTAATATTACGTAATTAATCCGGCCATAAAAATTACGTAATGGCCATAAAAGGCC
+ATAAAAGGCCATAAAAGGCCATAAAATAATCCATTACGTAATGCATAAAAAAATTACGTA
+ATATTACGTAATAA
+>11|1 This is a randomly generated sequence
+GGCCATAAAAGGCCATAAAAATTACGTAATATTACGTAATGGCCATAAAAATTACGTAAT
+GCATAAAAAAATTACGTAATGGCCATAAAAATTACGTAATATTACGTAATGGCCATAAAA
+TAATCCTAATCCAA
+>12|1 This is a randomly generated sequence
+TAATCCATTACGTAATATTACGTAATGGCCATAAAAATTACGTAATGCATAAAAAAATTA
+CGTAATGGCCATAAAAGGCCATAAAATAATCCATTACGTAATGGCCATAAAAATTACGTA
+ATGGCCATAAAAAA
+>13|0 This is a randomly generated sequence
+ATTACGTAATATTACGTAATATTACGTAATGGCCATAAAAGGCCATAAAAGGCCATAAAA
+ATTACGTAATGCATAAAAAATAATCCATTACGTAATGGCCATAAAATAATCCGGCCATAA
+AAATTACGTAATAA
+>14|0 This is a randomly generated sequence
+ATTACGTAATATTACGTAATTAATCCATTACGTAATGGCCATAAAAGGCCATAAAAATTA
+CGTAATGCATAAAAAAGGCCATAAAAGGCCATAAAATAATCCATTACGTAATATTACGTA
+ATGGCCATAAAAAA
+>15|0 This is a randomly generated sequence
+TAATCCGGCCATAAAAGGCCATAAAAATTACGTAATATTACGTAATATTACGTAATGGCC
+ATAAAAGGCCATAAAAGGCCATAAAAATTACGTAATATTACGTAATATTACGTAATTAAT
+CCGCATAAAAAAAA
+>16|0 This is a randomly generated sequence
+GGCCATAAAAGGCCATAAAAATTACGTAATATTACGTAATATTACGTAATGGCCATAAAA
+GCATAAAAAAGGCCATAAAAGGCCATAAAAATTACGTAATTAATCCATTACGTAATATTA
+CGTAATTAATCCAA
+>17|1 This is a randomly generated sequence
+GGCCATAAAAATTACGTAATGGCCATAAAAGCATAAAAAAGGCCATAAAAATTACGTAAT
+TAATCCGGCCATAAAAATTACGTAATATTACGTAATTAATCCATTACGTAATGGCCATAA
+AAATTACGTAATAA
+>18|1 This is a randomly generated sequence
+ATTACGTAATATTACGTAATATTACGTAATGGCCATAAAAATTACGTAATGGCCATAAAA
+GGCCATAAAATAATCCGCATAAAAAAATTACGTAATTAATCCATTACGTAATGGCCATAA
+AAGGCCATAAAAAA
+>19|0 This is a randomly generated sequence
+ATTACGTAATATTACGTAATGCATAAAAAAGGCCATAAAAGGCCATAAAAATTACGTAAT
+GGCCATAAAAGGCCATAAAATAATCCGGCCATAAAATAATCCATTACGTAATATTACGTA
+ATATTACGTAATAA
+>20|0 This is a randomly generated sequence
+ATTACGTAATGGCCATAAAATAATCCATTACGTAATTAATCCGGCCATAAAAGGCCATAA
+AAATTACGTAATATTACGTAATGGCCATAAAAATTACGTAATATTACGTAATGGCCATAA
+AAGCATAAAAAAAA
+>21|0 This is a randomly generated sequence
+TAATCCATTACGTAATGGCCATAAAAATTACGTAATTAATCCGGCCATAAAAGGCCATAA
+AAATTACGTAATATTACGTAATATTACGTAATGGCCATAAAAATTACGTAATGGCCATAA
+AAGCATAAAAAAAA
+>22|0 This is a randomly generated sequence
+ATTACGTAATGGCCATAAAATAATCCATTACGTAATATTACGTAATGGCCATAAAAATTA
+CGTAATATTACGTAATTAATCCGGCCATAAAAGCATAAAAAAGGCCATAAAAATTACGTA
+ATGGCCATAAAAAA
+>23|0 This is a randomly generated sequence
+ATTACGTAATATTACGTAATATTACGTAATGGCCATAAAAGGCCATAAAAGGCCATAAAA
+TAATCCGGCCATAAAAATTACGTAATATTACGTAATATTACGTAATGCATAAAAAATAAT
+CCGGCCATAAAAAA
+>24|1 This is a randomly generated sequence
+ATTACGTAATATTACGTAATGCATAAAAAAATTACGTAATATTACGTAATGGCCATAAAA
+ATTACGTAATTAATCCGGCCATAAAAGGCCATAAAAGGCCATAAAAGGCCATAAAATAAT
+CCATTACGTAATAA
