+++ example_outputs/neutralseq.fa
@@ -1,0 +1,2 @@
+>sequence
+ATTACGTAATATTACGTAATGCATAAAAAAATTACGTAATATTACGTAATGGCCATAAAAATTACGTAATTAATCCGGCCATAAAAGGCCATAAAAGGCCATAAAAGGCCATAAAATAATCCATTACGTAATAA
